perf: cheaper batch inserts: fewer storage overwrites
