+++ deployer/test/MerkleTreeController.test.js
@@ -8,18 +8,15 @@ import Web3 from '../src/web3';
 import deployer from './rest/deployer';
 
 const web3 = Web3.connect();
-const hashType = config.HASH_TYPE;
 
-let contractName = config.contractNames[0];
+const contractName = config.contractNames[0];
 let contractInstance;
 let coinbase;
 
-const numberOfBatches = 2;
-
-const batchSize = 5;
+const numberOfBatches = 1;
+const batchSize = 128;
 
 describe(`${contractName}`, async () => {
-
   before('get contractInstance', async () => {
     if (!(await Web3.isConnected())) await Web3.connection();
 
@@ -29,7 +26,7 @@ describe(`${contractName}`, async () => {
   });
 
   // eslint-disable-next-line func-names
-  describe(`adding leaves via ${contractName}`, async function() {
+  describe(`adding leaves (one-at-a-time) via ${contractName}`, async function() {
     this.timeout(3660000); // surprisingly, this.timeout() doesn't work inside an arrow function!
 
     // console.log('in config', config.contractNames, config.HASH_TYPE);
@@ -110,7 +107,7 @@ describe(`${contractName}`, async () => {
         ._insertLeaves(leaves)
         .send({
           from: coinbase,
-          gas: 8000000, // explore a full block of gas being used
+          gas: 10000000, // explore a full block of gas being used
           gasPrice: config.web3.options.defaultGasPrice,
         })
         // eslint-disable-next-line no-loop-func
@@ -124,7 +121,6 @@ describe(`${contractName}`, async () => {
 
       const { gasUsed } = txReceipt;
       gasUsedArray.push(gasUsed);
-      
     });
 
     after('provide summary stats', async () => {
@@ -138,61 +134,58 @@ describe(`${contractName}`, async () => {
       console.log('averageGasUsedMinusTxCost:', averageGasUsedMinusTxCost);
     });
   });
-
-
-
-
-  //eslint-disable-next-line func-names, prettier/prettier
-  describe(`Adding ${numberOfBatches * batchSize} leaves in batches of ${batchSize}`, async function() {
-    this.timeout(3660000); // surprisingly, this.timeout() doesn't work inside an arrow function!
-
-    const numberOfLeaves = numberOfBatches * batchSize;
-    const gasUsedArray = [];
-    let totalGasUsed = 0;
-    let averageGasUsed = 0;
-    let averageGasUsedMinusTxCost = 0;
-
-    it(`Adds the leaves`, async () => {
-      // create the leafValues to add:
-      const leaves = [];
-      for (let i = 0; i < numberOfLeaves; i += 1) {
-        const leaf = i.toString().padStart(64, '0'); // pad to 32 bytes
-        leaves.push(`0x${leaf}`);
-      }
-
-      for (let i = 0; i < numberOfBatches; i++) {
-        const leavesToInsert = leaves.slice(i * batchSize, (i + 1) * batchSize);
-        // eslint-disable-next-line no-await-in-loop
-        const txReceipt = await contractInstance.methods
-          ._insertLeaves(leavesToInsert)
-          .send({
-            from: coinbase,
-            gas: 8000000, // explore a full block of gas being used
-            gasPrice: config.web3.options.defaultGasPrice,
-          })
-          // eslint-disable-next-line no-loop-func
-          .on('receipt', receipt => {
-            const { minLeafIndex, leafValues, root } = receipt.events.NewLeaves.returnValues;
-
-            console.log('NewLeaves event returnValues:', minLeafIndex, leafValues, root);
-
-            // console.dir(receipt.events, { depth: null });
-          });
-
-        const { gasUsed } = txReceipt;
-        gasUsedArray.push(gasUsed);
-      }
-    });
-
-    after('provide summary stats', async () => {
-      totalGasUsed = gasUsedArray.reduce((acc, cur) => acc + cur);
-      averageGasUsed = totalGasUsed / batchSize;
-      averageGasUsedMinusTxCost = (totalGasUsed - 21000) / batchSize;
-      console.log('\ngasUsedArray:');
-      console.dir(gasUsedArray, { maxArrayLength: null });
-      console.log('totalGasUsed:', totalGasUsed);
-      console.log('averageGasUsed:', averageGasUsed);
-      console.log('averageGasUsedMinusTxCost:', averageGasUsedMinusTxCost);
-    });
-  });
+  //
+  // // eslint-disable-next-line func-names, prettier/prettier
+  // describe(`Adding ${numberOfBatches * batchSize} leaves in batches of ${batchSize}`, async function() {
+  //   this.timeout(3660000); // surprisingly, this.timeout() doesn't work inside an arrow function!
+  //
+  //   const numberOfLeaves = numberOfBatches * batchSize;
+  //   const gasUsedArray = [];
+  //   let totalGasUsed = 0;
+  //   let averageGasUsed = 0;
+  //   let averageGasUsedMinusTxCost = 0;
+  //
+  //   it(`Adds the leaves`, async () => {
+  //     // create the leafValues to add:
+  //     const leaves = [];
+  //     for (let i = 0; i < numberOfLeaves; i += 1) {
+  //       const leaf = i.toString().padStart(64, '0'); // pad to 32 bytes
+  //       leaves.push(`0x${leaf}`);
+  //     }
+  //
+  //     for (let i = 0; i < numberOfBatches; i++) {
+  //       const leavesToInsert = leaves.slice(i * batchSize, (i + 1) * batchSize);
+  //       // eslint-disable-next-line no-await-in-loop
+  //       const txReceipt = await contractInstance.methods
+  //         ._insertLeaves(leavesToInsert)
+  //         .send({
+  //           from: coinbase,
+  //           gas: 8000000, // explore a full block of gas being used
+  //           gasPrice: config.web3.options.defaultGasPrice,
+  //         })
+  //         // eslint-disable-next-line no-loop-func
+  //         .on('receipt', receipt => {
+  //           const { minLeafIndex, leafValues, root } = receipt.events.NewLeaves.returnValues;
+  //
+  //           console.log('NewLeaves event returnValues:', minLeafIndex, leafValues, root);
+  //
+  //           // console.dir(receipt.events, { depth: null });
+  //         });
+  //
+  //       const { gasUsed } = txReceipt;
+  //       gasUsedArray.push(gasUsed);
+  //     }
+  //   });
+  //
+  //   after('provide summary stats', async () => {
+  //     totalGasUsed = gasUsedArray.reduce((acc, cur) => acc + cur);
+  //     averageGasUsed = totalGasUsed / batchSize;
+  //     averageGasUsedMinusTxCost = (totalGasUsed - 21000) / batchSize;
+  //     console.log('\ngasUsedArray:');
+  //     console.dir(gasUsedArray, { maxArrayLength: null });
+  //     console.log('totalGasUsed:', totalGasUsed);
+  //     console.log('averageGasUsed:', averageGasUsed);
+  //     console.log('averageGasUsedMinusTxCost:', averageGasUsedMinusTxCost);
+  //   });
+  // });
 });
